chore(*): Create top-level workspace
Crates are pulled out of the vue root folder and a workspace is created
one level up.

I believe this approach is the most resilient for the future, as it
clearly separates the crates and lets the tauri-app be the gui only,
without mixing logic crates into it's root folder.
Having a workspace also makes it easier for IDE's to pick up the crates
and the compilation share's intermediate units.

Furthermore, I removed the tauri-vue-cli plugin - I had some, in the end
unrelated problems, regarding compilation and the tauri discord
server recommend me to remove it in general and use `@tauri-apps/cli`
directly, as the tauri-vue-cli plugin was not up-to-date anyway.
(@tauri-apps/cli is basically a wrapper around tauri-cli which can also
be installed and used via `cargo install tauri-cli`.)

Lsat but not least, I did not include the `libxl-test` crate in the
workspace - tests are a separate compilation unit and can be directly
part of the original crate so I am unsure of the need for that crate
(async functions can be tested with e.g. `#[tokio::test]`).

Ah, I also updated tauri and dependencies.